Use inheritance when prepending.
The PdfHelper module is prepended to ActionController::Base, but the
proper ruby prepend patter is not used. Instead of defining the method
to be overridden in the module and prepending that module in the target
class, the module has uses its self.prepended method to open up the
target class (with class_eval), and defines the method right there on
the target class (as well as aliasing the old render method). This
monkeypatching is exactly the sort of thing prepend is supposed to help
you get away from.

The monkeypatching can cause serious issue if any other gems are
prepending to ActionController::Base, since those gems will (hopefully)
use the new prepend pattern and call super to delegate to the original
implementation, but super now calls the version of render defined in
wicked's class_eval. That render then calls render_without_wicked_pdf,
which is now aliased to any render override that has been prepended
before wicked, and you get an infinite loop.